Allow to download with wget
Summary:
There is an HTTP redirection loop when file_get_contents, curl or
Guzzle is used to fetch a page of Le Soir website, but not with wget.

We so implement a fallback solution to download with this software.

Test Plan:
Tested with http://www.lesoir.be/1046282/article/actualite/regions/bruxelles/2015-11-17/un-sapin-21-metres-orne-grand-place-bruxelles-photos

Test provided.

Reviewers: xcombelle, dereckson

Projects: #security

Maniphest Tasks: T697

Differential Revision: http://devcentral.nasqueron.org/D307